requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
chromadb>=0.4.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
apache-airflow>=2.7.0
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
tiktoken>=0.5.0
tenacity>=8.2.0
pytest>=7.4.0
